migrate ConsumerReportClosedNoAction
