+++ app/services/emails/Email.scala
@@ -1,6 +1,5 @@
 package services.emails
 
-import models.company.Company
 import models.report.Report
 import models.EmailValidation
 import play.api.i18n.Lang
@@ -35,22 +34,6 @@ trait ConsumerEmail extends Email
 object Email {
 
   // ======= Conso =======
-
-  final case class ConsumerReportClosedNoAction(report: Report, maybeCompany: Option[Company], messagesApi: MessagesApi)
-      extends ConsumerEmail {
-    private val lang                                        = Lang(getLocaleOrDefault(report.lang))
-    implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
-
-    override val recipients: List[EmailAddress] = List(report.email)
-    override val subject: String                = messagesApi("ReportNotAnswered.subject")(lang)
-
-    override def getBody: (FrontRoute, EmailAddress) => String = (frontRoute, _) =>
-      views.html.mails.consumer.reportClosedByNoAction(report, maybeCompany)(frontRoute, messagesProvider).toString
-
-    override def getAttachements: AttachmentService => Seq[Attachment] =
-      _.needWorkflowSeqForWorkflowStepN(4, report)
-
-  }
 
   final case class ConsumerValidateEmail(
       emailValidation: EmailValidation,
