+++ app/services/emails/EmailDefinitionsConsumer.scala
@@ -387,6 +387,44 @@ object EmailDefinitionsConsumer {
 
   }
 
+  case object ConsumerReportClosedNoAction extends EmailDefinition {
+    override val category = Consumer
+
+    override def examples =
+      Seq(
+        "report_closed_no_action" -> ((recipient, messagesApi) =>
+          EmailImpl(genReport.copy(email = recipient), Some(genCompany), messagesApi)
+        ),
+        "report_closed_no_action_case_dispute" -> ((recipient, messagesApi) =>
+          EmailImpl(
+            genReport.copy(email = recipient, tags = List(ReportTag.LitigeContractuel)),
+            Some(genCompany),
+            messagesApi
+          )
+        )
+      )
+
+    final case class EmailImpl(
+        report: Report,
+        maybeCompany: Option[Company],
+        messagesApi: MessagesApi
+    ) extends ConsumerEmail {
+      private val lang                                        = Lang(getLocaleOrDefault(report.lang))
+      implicit private val messagesProvider: MessagesProvider = MessagesImpl(lang, messagesApi)
+
+      override val recipients: List[EmailAddress] = List(report.email)
+      override val subject: String                = messagesApi("ReportNotAnswered.subject")(lang)
+
+      override def getBody: (FrontRoute, EmailAddress) => String = (frontRoute, _) =>
+        views.html.mails.consumer.reportClosedByNoAction(report, maybeCompany)(frontRoute, messagesProvider).toString
+
+      override def getAttachements: AttachmentService => Seq[Attachment] =
+        _.needWorkflowSeqForWorkflowStepN(4, report)
+
+    }
+
+  }
+
   private def getLocaleOrDefault(locale: Option[Locale]): Locale = locale.getOrElse(Locale.FRENCH)
 
 }
